Refactor unittests
